refactor: move KeyCode extension methods and associated tests to their own files
